* separate client and server
