+++ example/eureka-server/src/main/resources/bootstrap.yml
@@ -1,0 +1,23 @@
+server:
+  port: 8761
+
+spring:
+  application:
+    name: eureka-server
+
+eureka:
+  instance:
+    hostname: localhost
+    lease-expiration-duration-in-seconds: 30
+    lease-renewal-interval-in-seconds: 30
+  client:
+    registerWithEureka: false
+    fetchRegistry: false
+    serviceUrl:
+            defaultZone: http://localhost:8761/eureka/
+  server:
+    enable-self-preservation: false
+
+endpoints:
+ shutdown:
+  enabled: true
