+++ example/grpc-client/build.gradle
@@ -1,12 +1,12 @@
 apply plugin: 'org.springframework.boot'
 
 dependencies {
-    compile('org.springframework.cloud:spring-cloud-starter-consul-discovery')
+    compile('org.springframework.cloud:spring-cloud-starter-eureka')
     compile('org.springframework.cloud:spring-cloud-starter-zipkin')
     compile('org.springframework.boot:spring-boot-starter-actuator')
 
     compile project(':example:grpc-lib')
-    compile project(':grpc-spring-boot-starter')
+    compile project(':grpc-client-spring-boot-starter')
 }
 
 buildscript {
