+++ example/grpc-client/src/main/resources/application.properties
@@ -1,3 +1,7 @@
 server.port=8888
 spring.application.name=grpc-client
 spring.sleuth.sampler.percentage=1
+eureka.instance.hostname=localhost
+eureka.client.registerWithEureka=true
+eureka.client.fetchRegistry=true
+eureka.client.serviceUrl.defaultZone=http://localhost:8761/eureka/
