+++ example/grpc-server/src/main/java/net/devh/grpc/GrpcServerApplication.java
@@ -1,19 +1,18 @@
 package net.devh.grpc;
 
-import net.devh.springboot.autoconfigure.grpc.server.EnableGrpcServer;
-
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
+import org.springframework.cloud.netflix.eureka.EnableEurekaClient;
 
 /**
  * User: Michael
  * Email: yidongnan@gmail.com
  * Date: 2016/11/8
  */
+@EnableEurekaClient
 @EnableDiscoveryClient
 @SpringBootApplication
-@EnableGrpcServer
 public class GrpcServerApplication {
 
     public static void main(String[] args) {
