+++ example/grpc-server/src/main/resources/application.properties
@@ -2,3 +2,7 @@ spring.application.name=grpc-server
 server.port=0
 grpc.server.port=9999
 spring.sleuth.sampler.percentage=1
+eureka.instance.hostname=localhost
+eureka.client.registerWithEureka=true
+eureka.client.fetchRegistry=true
+eureka.client.serviceUrl.defaultZone=http://localhost:8761/eureka/
