+++ grpc-client-spring-boot-autoconfigure/src/main/java/net/devh/springboot/autoconfigure/grpc/client/GlobalClientInterceptorConfigurerAdapter.java
