+++ grpc-client-spring-boot-autoconfigure/src/main/resources/META-INF/spring.factories
@@ -1,0 +1,3 @@
+# AutoConfiguration
+org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
+net.devh.springboot.autoconfigure.grpc.client.GrpcClientAutoConfiguration
