+++ grpc-server-spring-boot-autoconfigure/build.gradle
@@ -1,0 +1,19 @@
+apply plugin: 'propdeps'
+apply plugin: 'propdeps-maven'
+apply plugin: 'propdeps-idea'
+apply plugin: 'propdeps-eclipse'
+apply from: '../deploy.gradle'
+
+group = "net.devh"
+version = "${projectVersion}"
+
+compileJava.dependsOn(processResources)
+
+dependencies {
+    compileOnly("org.springframework.boot:spring-boot-configuration-processor:${springBootVersion}")
+    provided("org.springframework.boot:spring-boot-starter:${springBootVersion}")
+    provided("org.springframework.cloud:spring-cloud-starter-sleuth:${springSleuthVersion}")
+    provided("org.springframework.cloud:spring-cloud-starter-consul-discovery:${springCloudConsul}")
+    provided("org.springframework.cloud:spring-cloud-starter-eureka:${springCloudEureka}")
+    provided("io.grpc:grpc-all:${grpcVersion}")
+}
