+++ grpc-server-spring-boot-autoconfigure/src/main/resources/META-INF/spring.factories
@@ -1,0 +1,5 @@
+# AutoConfiguration
+org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
+net.devh.springboot.autoconfigure.grpc.server.GrpcServerAutoConfiguration,\
+net.devh.springboot.autoconfigure.grpc.server.GrpcMetedataEurekaConfiguration,\
+net.devh.springboot.autoconfigure.grpc.server.GrpcMetedataConsulConfiguration
