--- grpc-spring-boot-autoconfigure/src/main/java/net/devh/springboot/autoconfigure/grpc/client/TraceClientAutoConfiguration.java
@@ -1,33 +1,0 @@
-package net.devh.springboot.autoconfigure.grpc.client;
-
-import org.springframework.boot.autoconfigure.condition.ConditionalOnClass;
-import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
-import org.springframework.cloud.sleuth.Tracer;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.core.Ordered;
-import org.springframework.core.annotation.Order;
-
-/**
- * User: Michael
- * Email: yidongnan@gmail.com
- * Date: 5/17/16
- */
-@Configuration
-@ConditionalOnProperty(value = "spring.sleuth.scheduled.enabled", matchIfMissing = true)
-@ConditionalOnClass(Tracer.class)
-//@ConditionalOnBean(Tracer.class)
-@Order(Ordered.HIGHEST_PRECEDENCE)
-public class TraceClientAutoConfiguration {
-
-    @Bean
-    public GlobalClientInterceptorConfigurerAdapter globalTraceClientInterceptorConfigurerAdapter(Tracer tracer) {
-        return new GlobalClientInterceptorConfigurerAdapter() {
-
-            @Override
-            public void addClientInterceptors(GlobalClientInterceptorRegistry registry) {
-                registry.addClientInterceptors(new TraceClientInterceptor(tracer, new MetadataInjector()));
-            }
-        };
-    }
-}
